Songs router uses imported types and actions; created actions dir and file; converted pool to typescript
