--- server/routes/pool.js
@@ -1,6 +1,0 @@
-const { MongoClient } = require("mongodb");
-
-const uri = process.env.MONGODB_URI
-const client = new MongoClient(uri);
-
-module.exports = client;
